Ошибки вынес в папку
