+++ lib/API/exceptions/bad_request_exception.dart
@@ -1,0 +1,10 @@
+import 'package:dio/dio.dart';
+
+class BadRequestException extends DioError {
+  BadRequestException(RequestOptions r) : super(requestOptions: r);
+
+  @override
+  String toString() {
+    return 'Invalid request';
+  }
+}
