+++ lib/API/exceptions/conflict_exception.dart
@@ -1,0 +1,10 @@
+import 'package:dio/dio.dart';
+
+class ConflictException extends DioError {
+  ConflictException(RequestOptions r) : super(requestOptions: r);
+
+  @override
+  String toString() {
+    return 'Conflict occurred';
+  }
+}
