+++ lib/API/exceptions/deadline_exceeded_exception.dart
@@ -1,0 +1,10 @@
+import 'package:dio/dio.dart';
+
+class DeadlineExceededException extends DioError {
+  DeadlineExceededException(RequestOptions r) : super(requestOptions: r);
+
+  @override
+  String toString() {
+    return 'The connection has timed out, please try again.';
+  }
+}
