+++ lib/API/exceptions/internal_server_error_exception.dart
@@ -1,0 +1,10 @@
+import 'package:dio/dio.dart';
+
+class InternalServerErrorException extends DioError {
+  InternalServerErrorException(RequestOptions r) : super(requestOptions: r);
+
+  @override
+  String toString() {
+    return 'Unknown error occurred, please try again later.';
+  }
+}
