+++ lib/API/exceptions/no_internet_connection_exception.dart
@@ -1,0 +1,10 @@
+import 'package:dio/dio.dart';
+
+class NoInternetConnectionException extends DioError {
+  NoInternetConnectionException(RequestOptions r) : super(requestOptions: r);
+
+  @override
+  String toString() {
+    return 'No internet connection detected, please try again.';
+  }
+}
