+++ lib/API/exceptions/not_found_exception.dart
@@ -1,0 +1,10 @@
+import 'package:dio/dio.dart';
+
+class NotFoundException extends DioError {
+  NotFoundException(RequestOptions r) : super(requestOptions: r);
+
+  @override
+  String toString() {
+    return 'The requested information could not be found';
+  }
+}
