+++ lib/API/exceptions/unauthorized_exception.dart
@@ -1,0 +1,10 @@
+import 'package:dio/dio.dart';
+
+class UnauthorizedException extends DioError {
+  UnauthorizedException(RequestOptions r) : super(requestOptions: r);
+
+  @override
+  String toString() {
+    return 'Access denied';
+  }
+}
